Added `verifyReportAllowed()` to verify that a report is allowed for export
SQL is modeled from the report list check in `getProjectReports()`.  The `verifyReportAllowed()` function is derived from the `verifyExportRights()` function and should return a true/false.

Function is used in `getReport()` and `getReportColumns()`.

Also updated error message text for calls to `verifyExportRights()` to discriminate between these calls and calls to `verifyReportAllowed()`.